+++ utoipa/CHANGELOG.md
@@ -7,6 +7,7 @@ to look into changes introduced to **`utoipa-gen`**.
 
 ### Added
 
+* Also use generics on `ObjectBuilder::default(value)` like other setter (https://github.com/juhaku/utoipa/pull/1537)
 * Add support for jiff v0.2 `Timestamp` (https://github.com/juahku/utoipa/pull/1416)
 
 ## 5.4.0 - Jun 16 2025